Put builder different class #225
